remove const dupelication for smaller lybc files

+++ include/parser.h
@@ -52,7 +52,7 @@ typedef struct {
     int globalCount;
 
     Value *consts;
-    int constAmt;
+    uint64_t constAmt;
 
     Func *funcs;
     int funcAmt;
